requirements.txt
# python 3.10.0
django
assemblyai
openai
python-decouple
hugchat